refactor: Move transform mission logic into transformer
