+++ api/lib/infrastructure/transformers/index.js
@@ -3,3 +3,4 @@ export * as competenceTransformer from './competence-transformer.js';
 export * as skillTransformer from './skill-transformer.js';
 export * as tubeTransformer from './tube-transformer.js';
 export * as tutorialTransformer from './tutorial-transformer.js';
+export * as missionTransformer from './mission-transformer.js';
